Add etcd Role
etcd is added in the project as a key-value storage and a service
discovery mechanism for Calico. It is used to distribute information
about workloads, Calico endpoints, and policies.

etcd is started in a Docker container and managed by systemd on each
host for redundancy. Since we register etcd in Consul as a service, we can
use it for other purposes as well. Calico does not support other key-value
storages such as Consul, but we might want to extend it with such
functionality in the future.